+++ frontend-web/src/components/login/login-process-junction/view.tsx
@@ -1,0 +1,7 @@
+import LoadingSpinner from "@/components/loading-spinner/view";
+
+function LoginProcessJunctionPage() {
+  return <LoadingSpinner />;
+}
+
+export default LoginProcessJunctionPage;
