+++ frontend-web/src/components/login/terms/view.controller.ts
@@ -1,13 +1,10 @@
-import { useCallback, useEffect, useState } from "react";
-import { Cookies } from "react-cookie";
-import { logOnDev } from "@util/log-on-dev.ts";
+import { useCallback, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { checkIsAgreed, setIsOptionalAgreementAccepted } from "@api/auth/user/api.ts";
+import { setIsOptionalAgreementAccepted } from "@api/auth/user/api.ts";
 import { isAxiosError } from "axios";
 
 export function useTermsPageViewController() {
   /* React States */
-  const [isLoading, setIsLoading] = useState<boolean>(false);
   const [snackbarOpen, setSnackbarOpen] = useState<boolean>(false);
   const [isAgreed, setIsAgreed] = useState<boolean>(false);
 
@@ -15,14 +12,6 @@ export function useTermsPageViewController() {
   const navigate = useNavigate();
 
   /* Functions */
-  const redirectToSocialLoginNotYeyPage = useCallback(() => {
-    logOnDev(
-      `🚨 [Unauthorized Access] User Is About To Do Identification Without Social Login. Redirect To Failure Page...`,
-    );
-    navigate("/login/social/not-yet");
-    return;
-  }, [navigate]);
-
   const onClickIsAgreed = useCallback(() => {
     setIsAgreed(!isAgreed);
   }, [isAgreed]);
@@ -30,7 +19,7 @@ export function useTermsPageViewController() {
   const onClickNextButton = useCallback(() => {
     setIsOptionalAgreementAccepted(true)
       .then(() => {
-        navigate("/login/identify", { state: { fromTerms: true } });
+        navigate("/login/identify");
       })
       .catch((error) => {
         if (isAxiosError(error)) {
@@ -39,39 +28,5 @@ export function useTermsPageViewController() {
       });
   }, [navigate]);
 
-  /* useEffects */
-  useEffect(() => {
-    setIsLoading(true);
-
-    const cookies = new Cookies();
-    const accessToken = cookies.get("accessToken");
-
-    if (!accessToken || accessToken === "") {
-      redirectToSocialLoginNotYeyPage();
-      return;
-    }
-
-    cookies.remove("accessToken", { path: "/" });
-
-    checkIsAgreed()
-      .then((response) => {
-        const isAgreed = response.data.data.isOptionalAgreementAccepted;
-
-        if (isAgreed !== null) {
-          navigate("/login/identify", { state: { fromTerms: true } });
-          return;
-        }
-      })
-      .catch((error) => {
-        if (isAxiosError(error)) {
-          redirectToSocialLoginNotYeyPage();
-          return;
-        }
-      })
-      .finally(() => {
-        setIsLoading(false);
-      });
-  }, [navigate, redirectToSocialLoginNotYeyPage]);
-
-  return { isAgreed, onClickIsAgreed, onClickNextButton, snackbarOpen, setSnackbarOpen, isLoading };
+  return { isAgreed, onClickIsAgreed, onClickNextButton, snackbarOpen, setSnackbarOpen };
 }
